% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clusteredEnrichmentGO.R
\name{clusteredEnrichmentGO}
\alias{clusteredEnrichmentGO}
\title{clusteredEnrichmentGO}
\usage{
clusteredEnrichmentGO(gomatrix, geneset, universe, geneAlias = NULL,
  min.genes = 5, cut_max = 5, dynamic_cut = F, domains = c("BP", "MF",
  "CC"), two_sided = F, representative_by_silhouette = F,
  sort_by_pval = T, weights = NULL)
}
\arguments{
\item{gomatrix}{Sparse binary matrix, maps gene identifiers (column names) to the GO terms they are annotated with (row names).}

\item{geneset}{Character vector, gene set of interest.}

\item{universe}{Character vector, background set of genes. Note: gene identifiers not included in the gomatrix are dropped!}

\item{geneAlias}{Named character vector, maps gene identifiers (vector names) to alternative identifiers.}

\item{min.genes}{Minimum number of genes of interest that a term needs to contain to be included in the analysis.}

\item{cut_max}{Terms may be combined into a cluster if they contain at most cut_max different genes of interest (i.e. max. edit distance).}

\item{dynamic_cut}{Default: FALSE. If TRUE, evaluate the silhouette index of all clusterings with max. edit distances of [1...cut_max], and choose the optimal cut accordingly.}

\item{domains}{Character vector, select included GO terms by domain: biological process ("BP"), cellular compartment ("CC"), molecular function ("MF").}

\item{two_sided}{Default: FALSE. If TRUE, performs two-sided Fisher tests to detect depletion as well as enrichment.}

\item{representative_by_silhouette}{Default: FALSE. If TRUE, selects the representative of a cluster based on their silhouette score 
(i.e. the most specific term for that cluster). Otherwise, the smallest term is chosen, or the term with the shortest description in the presence of ties.}

\item{sort_by_pval}{Default: TRUE. Sort results by p-value.}
}
\description{
Gene Ontology enrichment with subsequent clustering of terms
}
\examples{
\dontrun{
# example for annotation matrix creation
# may be used interchangably with createGoMatrix()
gomatrix = createAnnotationMatrix( list("GO:0000042"=paste0("gene",1:6),
                                           "GO:0000002"=paste0("gene",2:7),
                                           "GO:0000009"=paste0("gene",6:12)) )

# example call for main function
res = clusteredEnrichmentGO(gomatrix, 
                            geneset = paste0("gene",2:6),
                            universe = paste0("gene",1:12))

# example plot
plotClusteredEnrichment(res$results)
}

}
\author{
Robert Sehlke
}
