% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createGoMatrix.R
\name{createGoMatrix}
\alias{createGoMatrix}
\title{createGoMatrix}
\usage{
createGoMatrix(egGO2ALLEGS, min.genes = 5, max.genes = 300,
  excludeIEA = F)
}
\arguments{
\item{egGO2ALLEGS}{GO to ALL identifiers mapping from any bioconductor annotation package. E.g. org.Dm.egGO2ALLEGS from the
org.Dm.eg.db package.}

\item{min.genes}{Minimum size of gene set to be considered, defaults to 5.}

\item{max.genes}{Maximum size of gene set to be considered, defaults to 300.}

\item{excludeIEA}{}
}
\description{
Using a bioconductor gene ontology annotation package, create a sparse binary matrix that maps gene identifiers (column names) to the 
GO terms they are annotated with (row names).
}
\examples{
\dontrun{
# example for annotation matrix creation
# may be used interchangably with createGoMatrix()
gomatrix = createAnnotationMatrix( list("GO:0000042"=paste0("gene",1:6),
                                           "GO:0000002"=paste0("gene",2:7),
                                           "GO:0000009"=paste0("gene",6:12)) )

# example call for main function
res = clusteredEnrichmentGO(gomatrix, 
                            geneset = paste0("gene",2:6),
                            universe = paste0("gene",1:12))

# example plot
plotClusteredEnrichment(res$results)
}

}
\author{
Robert Sehlke
}
