% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotClusteredEnrichment.R
\name{plotClusteredEnrichment}
\alias{plotClusteredEnrichment}
\title{plotClusteredEnrichment}
\usage{
plotClusteredEnrichment(ceresult, alpha = 0.05, t.alpha = 0.05,
  max_bars = NULL, top_n = Inf, only_positive_enrichment = T,
  order_by_enrichment = T, main = "Enriched primary gene sets",
  legend.title = "", xlim = NULL, lab_max_chars = 50, ...)
}
\arguments{
\item{ceresult}{$results data frame returned by the clusteredEnrichmentGO() function.}

\item{alpha}{Significance cutoff for primary gene set adjusted p-values, defaults to 0.05.}

\item{t.alpha}{If a directionality bias has been determined via t-test on the log-fold-change of the gene set,
which significance cutoff should be used for the display of that bias? (i.e. bar color)}

\item{max_bars}{If set, scales the bar plot to provide space for max_bars... bars. Use to ensure uniform scaling 
when plotting multiple results with varying numbers of significant hits.}

\item{top_n}{Plot only the most significant top_n results. Default is to show all results.}

\item{only_positive_enrichment}{Defaults to TRUE: show only positive log-enrichments (i.e. no depleted terms).}

\item{order_by_enrichment}{Defaults to TRUE: order significant hits by effect magnitude (enrichment) rather than
significance.}

\item{legend.title}{Title of the legend (only relevant if directionality bias of GO terms is included in the results).}

\item{xlim}{Axis limit of the enrichment plot. Use to ensure uniform axis across multiple plots.}

\item{lab_max_chars}{Maximum length of gene set names; longer names are truncated.}
}
\description{
Simple barplot that displays the enrichment of primary gene sets from a clusteredEnrichment result.
}
\examples{
\dontrun{
# example for annotation matrix creation
# may be used interchangably with createGoMatrix()
gomatrix = createAnnotationMatrix( list("GO:0000042"=paste0("gene",1:6),
                                           "GO:0000002"=paste0("gene",2:7),
                                           "GO:0000009"=paste0("gene",6:12)) )

# example call for main function
res = clusteredEnrichmentGO(gomatrix, 
                            geneset = paste0("gene",2:6),
                            universe = paste0("gene",1:12))

# example plot
plotClusteredEnrichment(res$results)
}

}
\author{
Robert Sehlke
}
